Refactor flag creation and rename simple -> simple_boolean_flag
