+++ docs/api_docs/bundle.yaml
@@ -864,7 +864,7 @@ definitions:
         description: unique key representation of the flag
         type: string
       template:
-        description: template name for creation
+        description: template for flag creation
         type: string
   putFlagRequest:
     type: object
