+++ pkg/handler/crud.go
@@ -114,86 +114,6 @@ func (c *crud) FindFlags(params flag.FindFlagsParams) middleware.Responder {
 	return resp
 }
 
-func (c *crud) CreateFlag(params flag.CreateFlagParams) middleware.Responder {
-	f := &entity.Flag{}
-	if params.Body != nil {
-		f.Description = util.SafeString(params.Body.Description)
-		f.CreatedBy = getSubjectFromRequest(params.HTTPRequest)
-
-		key, err := entity.CreateFlagKey(params.Body.Key)
-		if err != nil {
-			return flag.NewCreateFlagDefault(400).WithPayload(
-				ErrorMessage("cannot create flag. %s", err))
-		}
-		f.Key = key
-	}
-
-	tx := getDB().Begin()
-
-	if err := tx.Create(f).Error; err != nil {
-		tx.Rollback()
-		return flag.NewCreateFlagDefault(500).WithPayload(
-			ErrorMessage("cannot create flag. %s", err))
-	}
-
-	if params.Body.Template == "simple" {
-		// Create our default segment
-		s := &entity.Segment{}
-		s.FlagID = f.ID
-		s.RolloutPercent = uint(100)
-		s.Rank = entity.SegmentDefaultRank
-
-		if err := tx.Create(s).Error; err != nil {
-			tx.Rollback()
-			return flag.NewCreateFlagDefault(500).WithPayload(ErrorMessage("%s", err))
-		}
-
-		// .. and our default Variant
-		v := &entity.Variant{}
-		v.FlagID = f.ID
-		v.Key = "on"
-
-		if err := tx.Create(v).Error; err != nil {
-			tx.Rollback()
-			return flag.NewCreateFlagDefault(500).WithPayload(ErrorMessage("%s", err))
-		}
-
-		// .. and our default Distribution
-		d := &entity.Distribution{}
-		d.SegmentID = s.ID
-		d.VariantID = v.ID
-		d.VariantKey = v.Key
-		d.Percent = uint(100)
-
-		if err := tx.Create(d).Error; err != nil {
-			tx.Rollback()
-			return flag.NewCreateFlagDefault(500).WithPayload(ErrorMessage("%s", err))
-		}
-
-		s.Distributions = append(s.Distributions, *d)
-		f.Variants = append(f.Variants, *v)
-		f.Segments = append(f.Segments, *s)
-	}
-
-	err := tx.Commit().Error
-	if err != nil {
-		tx.Rollback()
-		return flag.NewCreateFlagDefault(500).WithPayload(ErrorMessage("%s", err))
-	}
-
-	resp := flag.NewCreateFlagOK()
-	payload, err := e2rMapFlag(f)
-	if err != nil {
-		return flag.NewCreateFlagDefault(500).WithPayload(
-			ErrorMessage("cannot map flag. %s", err))
-	}
-	resp.SetPayload(payload)
-
-	entity.SaveFlagSnapshot(getDB(), f.ID, getSubjectFromRequest(params.HTTPRequest))
-
-	return resp
-}
-
 func (c *crud) GetFlag(params flag.GetFlagParams) middleware.Responder {
 	f := &entity.Flag{}
 	result := entity.PreloadSegmentsVariants(getDB()).First(f, params.FlagID)
